content/media/video/test/test_ended2.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: ended and replaying</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v'
       src='seek.ogv'
       onloadedmetadata='return startTest();'
       onplay='return playbackStarted();'
       onended='return playbackEnded();'></video>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test if video can be replayed after ended.
var v = document.getElementById('v');
var completed = false;
var playCount = 0;

function startTest() {
  if (completed)
    return false;

  v.play();
  return false;
}

function playbackStarted() {
  if (completed)
    return false;

  playCount++;
}

function playbackEnded() {
  if (completed)
    return false

  ok(v.currentTime >= 3.9 && v.currentTime <= 4.0,
     "Checking currentTime at end: " + v.currentTime);
  ok(v.ended, "Checking playback has ended");
  if (playCount < 2) {
    v.play();
  }
  else {
    ok(playCount == 2, "Check playback after ended event");
    completed = true;
    SimpleTest.finish();
  }
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
